refactor: make the network shim cypress-plugin based (#128)
* refactor: make the network shim cypress-plugin based
* chore(network-shim): remove stale files
* fix(network-shim): throw error when detecting parallel capture run
* fix(network-shim): stop relying on aliases across before and beforeEach (This has proven to be unreliable)
* chore(network-shim): regenerate corrupt fixture files
* fix(network-shim): add networkshim plugin via install command

BREAKING CHANGE: To use the network-shim's functionality the project will need to use the new network shim plugin on top of the existing enableNetworkShim setup. Also, the network shim options (an object with hosts and staticResources) need to be passed to the networkShim plugin, the enableNetworkShim setup doesn't accept any arguments anymore.