fix(network-shim): ensure in-test fixtures are used instead of shim fixtures
The core issue was fixed by upgrading cypress to v7, which includes the correct overriding behaviour for `cy.intercept`.
Due to the major version bump, capturing and stubbing requests stopped working, and this was fixed by passing regexes to cy.intercept instead of strings.
A new test was introduced to the network-shim example-app to assert the in-test fixtures take presidence over the networkshim fixtures.

BREAKING CHANGE: bumps cypress 1 major version, to v7