refactor: wait for baseUrl to become available
Wrap the Cypress command in wait-on if we can find a baseUrl from the
project's cypress.json file. If no baseUrl or cypress configuration can
be found, we just run the command as is.

BREAKING CHANGE: Drop the --waitOn flag
As of now we wait on the baseUrl that is defined in cypress.json, as
that is the URL that the tests are going to run against.