refactor: simplify cypress-plugin and cli-utils-cypress
Refactor the way the way d2-utils-cypress relates to cypress, and how it
is intended to be used in a project.

The biggest change is that a consumer does not use d2-utils-cypress to
set up the environment for Cypress and then automatically run Cypress
with those changes.

Instead, the consumer runs Cypress directly.

    cypress run

After initial configuration in either cypress.json through the "env"
key, or in cypress.env.json, the user can activate the plugin through:

    cypress run --env networkMode=capture
    cypress run --env networkMode=stub
    cypress run --env networkMode=live

"live" is the default if the network env property is not set, so it can
be omitted, but for clarity it is good practice to add it to the
cypress.env.json:

    {
        "dhis2Username": "admin",
        "dhis2Password": "district",
        "dhis2BaseUrl": "http://localhost:8080",
        "dhis2DataTestPrefix": "networkshim-platformapp",
    +   "dhis2ApiVersion": "37",
    +   "networkMode": "live"
    }

The cypress-plugin uses configuration key "dhis2ApiVersion" to switch
the expected API version of the DHIS2 server. A default can be supplied
in a config file as above, or overridden per run. For example a stub run
against version 36:

    cypress run --env networkMode=stub,dhis2ApiVersion=36

To separate out what is considered Cypress configuration and what is
considered environment variables, this changes the keys to align with
Cypress recommendations.

The configuration options for our plugin are intended to be passed in
with:

    cypress run --env networkMode=capture,dhis2BaseUrl=http://..

This makes them available to the browser code through Cypress.env().

The browser does not have access to the process.env, so that is where
the separation between Cypress environment variables and the OS-level
environment variables comes in to play.

We can add our configuration through the --env flag, to the env key in
cypress.json, or to cypress.env.json, as they will be loaded into the
Cypress.env().

This is the recommended way to do it.

In CI, using the --env flag is not very ergonomic, as it must be on one
line and can be quite long and hard to read.

The official workaround for this scenario is to use the 'CYPRESS_'
prefix for the configuration key names to signal to Cypress that it
should load these OS-level environment variables into Cypress.env().

So in our workflows it would look like:

    env:
      CYPRESS_dhis2BaseUrl: http
      CYPRESS_dhis2Username: foo
      CYPRESS_dhis2Password: bar
      CYPRESS_networkMode: capture

This is admittedly a learned pattern, but once it is known then it is
quite clear what is intended as OS-level environment, and what is
intended for the Cypress environment as configuration.

As a bonus this commit extracts the isStubMode, isLiveMode,
isCaptureMode, getDhis2BaseUrl, and setDhis2BaseUrlToLocalStorage, into
helpers that are exposed through @dhis2/cypress-commands.

BREAKING CHANGE: Drop run and open commands
We want to be consistent with how Cypress runs locally and in CI and
since we cannot use d2-utils-cypress in CI, we shouldn't run it through
d2-utils-cypress locally either.

BREAKING CHANGE: Change configuration keys to camelCase.
- dhis2_username => dhis2Username
- dhis2_password => dhis2Password
- dhis2_base_url => dhis2BaseUrl
- dhis2_datatest_prefix => dhis2DataTestPrefix
- dhis2_api_version => dhis2ApiVersion

BREAKING CHANGE: dhis2_api_stub_mode renamed to networkMode
Instead of describing the mode of the plugin, it is a bit easier to
understand if we speak in terms of the network:
- do we want to capture the network traffic (networkMode=capture),
- do we want to stub it (networkMode=stub),
- or do we want to run it against a live backend (networkMode=live)?

BREAKING CHANGE: 'DISABLED' renamed to 'LIVE'
To better describe the state of the network when running tests instead
of describing the state of the plugin, DISABLED is now LIVE.

BREAKING CHANGE: isDisabledMode renamed to isLiveMode.
Similar to the above, to better describe the state of the network vs.
the state of the plugin, replace usages of isDisabledMode with
isLiveMode.

BREAKING CHANGE: 'CAPTURE'|'STUB'|'LIVE' are now lowercase when passed
to the environment.
Replace networkMode=LIVE with networkMode=live.