fix: remove uneducational & confusing install cli tool
BREAKING CHANGE: This removes the `d2-utils-cypress` cli tool!

Reason for this being a fix, see: https://stackoverflow.com/a/55213921/1319140
I'd agree that this is actually removing something we could consider a
"conceptual bug", because this kind of stuff shouldn't really be there
in the first place.

Removing the code that's been deleted allows us to better educate
developers how to use cypress the way we thing it works best within our
environment without hiding/wrapping any of cypress' functionality and
instead making developers learn how to use vanilla cypress.