feat: improve login by api command
The command no longer relies on a seemingly unrelated endpoint to infer
the correct major version of the backend. The amount of requests needed
to login is also reduced because the correct login mechanism is
persisted as a Cypress env var. And finally, the command will now always
yield a response, which can be asserted.